Transfer functionality Fixes #261

+++ mono/finance/signals.py
@@ -1,4 +1,4 @@
-from .models import Configuration, Account, Category, Group, Icon, Installment, Transaction, User
+from .models import Configuration, Account, Category, Group, Icon, Installment, Transaction, Transference, User
 
 
 def initial_setup(sender, instance, created, **kwargs):
@@ -56,6 +56,17 @@ def installments_creation(sender, instance, created, **kwargs):
             instance.create_transactions()
         else:
             transactions = Transaction.objects.filter(installment=instance)
+            for transaction in transactions:
+                transaction.delete()
+            instance.create_transactions()
+
+
+def transference_creation(sender, instance, created, **kwargs):
+    if sender == Transference:
+        if created:
+            instance.create_transactions()
+        else:
+            transactions = Transaction.objects.filter(transference=instance)
             for transaction in transactions:
                 transaction.delete()
             instance.create_transactions()
